// cargodelivery-front/generated/models/Settlement.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cargo delivery
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Point } from './Point';
import {
    PointFromJSON,
    PointFromJSONTyped,
    PointToJSON,
} from './Point';

/**
 * 
 * @export
 * @interface Settlement
 */
export interface Settlement {
    /**
     * 
     * @type {string}
     * @memberof Settlement
     */
    name: string;
    /**
     * 
     * @type {Point}
     * @memberof Settlement
     */
    coordinates: Point;
    /**
     * 
     * @type {string}
     * @memberof Settlement
     */
    id: string;
}

/**
 * Check if a given object implements the Settlement interface.
 */
export function instanceOfSettlement(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "coordinates" in value;
    isInstance = isInstance && "id" in value;

    return isInstance;
}

export function SettlementFromJSON(json: any): Settlement {
    return SettlementFromJSONTyped(json, false);
}

export function SettlementFromJSONTyped(json: any, ignoreDiscriminator: boolean): Settlement {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'coordinates': PointFromJSON(json['coordinates']),
        'id': json['id'],
    };
}

export function SettlementToJSON(value?: Settlement | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'coordinates': PointToJSON(value.coordinates),
        'id': value.id,
    };
}